+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added nodeID and address to operation and transport errors 
 * Prevented create decoder instance until start read a message from topics
 
 ## v3.99.4